Add a new "tags" collection, and use it to build tags
This update creates a new "tags" collection, where there's a meta-page
named "_tags/_base_tag.md". To activate a tag (have jekyll create a page
for it), all that is needed is to symlink the tags/<tag>.md to this meta
page. A helper script named _gen_tags.py exists to help discover and
create these symlinks.

On the front page, I dynamically generate a "tag cloud" that is
derived from the "explore_topics" array defined statically in
_config.yaml. Again, I provide a _get-tags.sh script that helps to
discover and generate this list semi-automatically.

Signed-off-by: Coleman Kane <ckane@colemankane.org>